Move functions to myGardenHelpers to improve code readability
