+++ src/routes/MyGarden.jsx
@@ -7,6 +7,12 @@ import PlantCard from "../components/PlantCard";
 import PlantModal from "../components/PlantModal";
 import { Link } from "react-router-dom";
 import { calculateNextWaterDate, dateFormatter } from "../helpers/dateHelpers";
+import {
+  performSearchPlant,
+  sortPlants,
+  waterSinglePlant,
+  waterAllPlants,
+} from "../helpers/myGardenHelpers";
 
 export default function MyGarden() {
   const { userID } = useContext(userContext);
@@ -16,58 +22,13 @@ export default function MyGarden() {
   const [plantCardProps, setPlantCardProps] = useState();
   const [filterPlants, setFilterPlants] = useState("needs water");
 
-  const searchPlant = (event) => {
-    event.preventDefault();
-    //Case insensitive search for plants that have the typed letter(s) in their generic or specific name
-    setSelectedPlants(
-      gardenInfo.filter((plant) =>
-        plant.specific_name
-          .toLowerCase()
-          .includes(event.target.value.toLowerCase())
-      )
-    );
-  };
-
-  const filterPlantsThatNeedWater = () =>
-    setSelectedPlants(
-      gardenInfo.filter((plant) => plant.waterStatus === "needs water")
-    );
-
-  //Updates the STATE of a plant that was watered in the LargeCardUser (LargeCardUser handles updating the database)
-  const waterSinglePlant = (id) => {
-    const plantsArray = [...selectedPlants];
-    plantsArray.forEach((plant) => {
-      const nextWateringCalc = calculateNextWaterDate(
-        new Date(),
-        plant.water_needs
-      );
-      if (plant.key_id === id) {
-        plant.lastWateredFormatted = dateFormatter(new Date());
-        plant.waterStatus = "watered";
-        plant.nextWaterFormatted = dateFormatter(nextWateringCalc);
-        plant.last_watered_at = new Date();
-        // plant.nextWatering = nextWateringCalc;
-        console.log(plant);
-      }
-    });
-    setSelectedPlants(plantsArray);
-  };
-
   const generateCards = () => {
     //Ensures there is data in gardenInfo
     if (gardenInfo[0]) {
-      //Sorts the plants alphabetically for display
-      selectedPlants.sort((a, b) => {
-        let lowerCaseA = a.specific_name.toLowerCase();
-        let lowerCaseB = b.specific_name.toLowerCase();
-
-        if (lowerCaseA < lowerCaseB) return -1;
-        if (lowerCaseA > lowerCaseB) return 1;
-        return 0;
-      });
-
+      //Sort plants alphabetically for display
+      const sortedPlants = sortPlants(selectedPlants);
       //Creates an array of PlantCards with the corresponding information
-      return selectedPlants.map((plant) => (
+      return sortedPlants.map((plant) => (
         <PlantCard
           key={plant.key_id}
           plant={plant.specific_name}
@@ -81,7 +42,7 @@ export default function MyGarden() {
               id: plant.key_id,
               waterStatus: plant.waterStatus,
               nextWatering: plant.nextWaterFormatted,
-              updateMyGarden: () => waterSinglePlant(plant.key_id),
+              updateMyGarden: () => waterSinglePlant(plant.key_id, gardenInfo),
             });
           }}
         />
@@ -114,27 +75,24 @@ export default function MyGarden() {
     });
   }, [userID]);
 
-  const handleWaterAllPlants = () => {
+  const searchPlant = (event) => {
+    event.preventDefault();
+    const searchTerm = event.target.value.toLowerCase();
+    setSelectedPlants(performSearchPlant(searchTerm, gardenInfo));
+  };
+
+  const filterPlantsThatNeedWater = () =>
+    setSelectedPlants(
+      gardenInfo.filter((plant) => plant.waterStatus === "needs water")
+    );
+
+  const handleWaterAllPlants = (plantsList) => {
     //Extract the plant_id of the plants that are visible at the time
     const idArray = selectedPlants.map((plant) => plant.key_id);
     //Update database, but only the plants that were visible at the time
     axios.put(`/api/my_garden/waterAll/${idArray}`);
     //Update state with the corresponding data
-    const today = new Date();
-    const wateredPlants = [...selectedPlants];
-    wateredPlants.forEach(
-      (plant) => (
-        (plant.lastWateredFormatted = dateFormatter(today)),
-        (plant.last_watered_at = today),
-        (plant.waterStatus = "watered"),
-        (plant.nextWatering = calculateNextWaterDate(
-          plant.last_watered_at,
-          plant.water_needs
-        )),
-        (plant.nextWaterFormatted = dateFormatter(plant.nextWatering))
-      )
-    );
-    setSelectedPlants(wateredPlants);
+    setSelectedPlants(waterAllPlants(plantsList));
   };
 
   const handleFilterPlants = () => {
@@ -171,7 +129,7 @@ export default function MyGarden() {
           className="col-3 offset-1"
           variant="primary"
           onClick={() => {
-            handleWaterAllPlants();
+            handleWaterAllPlants(selectedPlants);
           }}
         >
           Water All Plants
